pyproject.toml
[tool.poetry]
name = "repl_python3_nwHacks-2022"
version = "0.1.0"
description = ""
authors = ["DonaldKLee <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
pymongo = "^4.0.1"
dnspython = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
